ggscreeplot {ggbiplot}	R Documentation

Screeplot for Principal Components

Description

Screeplot for Principal Components

Usage


  ggscreeplot(pcobj, type = c("pev", "cev"))

Arguments

pcobj
an object returned by prcomp() or princomp()

type
the type of scree plot. 'pev' corresponds proportion of explained variance, i.e. the eigenvalues divided by the trace. 'cev' corresponds to the cumulative proportion of explained variance, i.e. the partial sum of the first k eigenvalues divided by the trace.

Examples


data(wine)
wine.pca <- prcomp(wine, scale. = TRUE)
print(ggscreeplot(wine.pca))

[Package ggbiplot version 0.55 Index]